Stop popups from making the page scrollable
The popup element can technically be partly outside the page, even though visually it is not. This causes the page to become horizontally scrollable, which can lead to problems like kiln/flourish#1267.

This commit adds an absolutely-positioned “constrainer” element containing the popup that has overlow:hidden, which covers the entire document. It’s surprisingly hard to achieve that – or at least I failed to find an easy way – and so we explicitly set the dimensions of the constrainer to match the document, and recompute them on window resize.

Fixes #16.