More robust fix for #16
@timbrock pointed out a potential problem with the previous implementation,
that some templates have resize handlers that change the DOM in response to
resize.

Also @micycle found that some templates have `body { height: 100% }` in the
CSS, which played havoc with the method we were using to work out the height
of the document. We now use the scroll height instead.

This new implementation adds only a single “constrainer” element to the document,
regardless of how many popups are on the page.