Improve performance of repeated calls to popup.hide()
Previously, _resizeConstrainer() was being called every time
popup.hide() was being called. This introduces a simple check
that prevents popup.hide() from running any heavy code if it's
already hidden.

This improves the performance of the popup module in the
point map template, which currently calls popup.hide() once
per frame. It's possible to make changes in the template to
prevent this, but I figure it's better to address this at the
module level to prevent it causing performance issues in other
templates in the future :)

See also:
https://github.com/kiln/template-3d-point-map/issues/97#issuecomment-662986017